packed-backend: add check for object consistency
If there is nothing wrong when parsing the raw file "packed-refs", we
could then iterate the "entries" to check the object consistency. There
are two kinds of ref entry: one is the normal and another is peeled. For
both situations, we need to use "parse_object" function to parse the
object id to get the object. If the object does not exist, we will
report an error to the user.

Create a new function "packed_fsck_ref_oid" to do above then update the
unit test to exercise the code.

Mentored-by: Patrick Steinhardt <ps@pks.im>
Mentored-by: Karthik Nayak <karthik.188@gmail.com>
Signed-off-by: shejialuo <shejialuo@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>